вынес photoUploader в отдельный файл
